Be less picky about detecting breakpoint ack
When setting a lot of breakpoints, occasionally the response comes
back prefixed by a normal prompt instead of subprompt, meaning
plain '(gdb)' instead of '>>>(gdb)' --  this did not get stripped
by the subprompt pattern, leading to the breakpoint acknowledgement
being thrown away as junk.

Simple solution: relax the breakpoint ack detection to ignore
any leading junk. The breakpoint ack is a pretty strong pattern in
itself, so it's not like there should be any false positives,
and the volume of text being matched is not sufficient to care about
match() vs search().